MB-39594: Test for 'Default' collection with 1 byte load/compare
Profiling a collection enabled workload showed functions taking a higher
percentage compared to a non-collection workload. This MB aims to reduce
the following:

* update_topkeys

update_topkeys wants to know if the key belongs to the default
collection. This is done by decoding the entire collection-ID and
comparing against Default (which is 0). This per operation cost can be
reduced as there is no need to do the full decode when testing for
values <= 127, byte 0 of the key can be inspected directly.

Change-Id: Ibabcd01f5095206fec306bc53786c0e9aa298fc7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131444
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>