MB-30524: Add collections support to stats key and vkey
Add collections support to stats key and vkey. Do this refactor
doKeyStats() and doVKeyStats() of EventuallyPersistentEngine to ensure
that a key/vkey stats request checks for a vbucket-id rather than
defaulting to vbucket-id 0. Also expand stats key/vkey to allow the
specification of a collection in the following formats.

key stats:
  key the-key vbid collection-path e.g. key mykey 0 _default._default
  key-byid the-key vbid collection-path e.g. key-byid mykey 0 0x0
vkey stats:
  vkey the-key vbid collection-path e.g. key mykey 0 _default._default
  vkey-byid the-key vbid collection-path e.g. key-byid mykey 0 0x0

Also update cbstats.py to be able to take advantage of these new
collection interfaces.

Change-Id: Ib6e4ab602646919bc16935600c6198d75787a148
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131745
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>