MB-40162: Allow Add+CreateAsDeleted if no tombstone found after bgfetch
When attempting to create a document directly in the Tombstoned state
(via CreateAsDeleted), this request would intermittently fail with
KEY_ENOENT.

This is caused by a bug in the subdoc_fetch code when allowing
execution to continue even if no existing doc is found (say when using
CreateAsDeleted), if the check to see if a tombstone is present needs
to go to disk (i.e. the Bloom filter failed to avoid the
bgFetch). This explains why this only occurs intermittently.

In that situation, the subdoc state machine sequence for a
multi-mutation with Add|AccessDeleted|CreateAsDeleted looks like:

1. subdoc_fetch -> bucket_get(AliveOrDeleted) ->
   EventuallyPersistentEngine::get() -> not found in HT and bloom
   filter cannot determine -> EWOULDBLOCK, schedule bgFetch.

2. subdoc_fetch returns back to runloop with EWOULDBLOCK, waiting for
   notify_io_complete

3. bgFetch completes, returns KEY_ENOENT (no tombstone on disk) ->
   notify_IO_complete(fd, KEY_ENOENT).

4. subdoc_fetch called again, checks AsyncIO status code
   KEY_ENOENT. CreateAsDeleted specified - Ok, setup empty document
   and return true (to continue execution).

*** However, the status code ret) is not modified, is still KEY_ENOENT. ***

5. subdoc_execute called, performs requested operations on
   newly-formed empty doc.

6. subdoc_update called, checks ret. Upon finding non-successful
result returns early with that same status code (KEY_ENOENT).

Note the problem at step 4. Fix is to ensure that after subdoc_fetch()
succeeds (i.e. is happy with no tombstone being present), then ret is
set to SUCCESS.

Change-Id: I6edb5df9f4cdbf986971136c1a3aa4c493c83711
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131460
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>