MB-40126: subdoc CreateAsDeleted: User value shouldn't be '{}'
When using the subdoc API to create docuements in the deleted state
with user XAttrs, the user value _should_ be entirely empty (apart
from the XATTRs).

However, the user value can incorrectly end up as '{}' (i.e. an empty
JSON object, but _not_ an empty document) in the following situation:

1. If the document doesn't already exist (but the user specifies
   doc_flag::MkDoc or doc_flag::Add).
2. If the subdoc mutation doesn't specify a path to the user value.

This occurs because if a existing document doens't exist (but MkDoc or
Add is specified), then a "empty" JSON object '{}' is assigned to be
the input document. If no futher modification of the document occurs,
then the resulting user body is empty JSON oject '{}'.

The solution is to defer creating the empty template object until the
body phase of subdoc_update; only if there's actually one or more
value paths to be applied.

Change-Id: I466d38ef71bd0345e4c45905d85ebfbb8bda55b6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131577
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>