MB-39297: Disable counting of prepares for magma
As magma is an LSM tree store it can have multiple version of a
given key. During compaction, the compaction callback can be called
multiple times for different versions of the same key. This is an
issue as we attempt to adjust the prepare count (and the document
count) as we visit each prepare.

There does exist a solution to this where we can bg fetch the prepare
to see if it is the most recent during a compaction. This would require
a bunch of code though as we couldn't have the bgfetcher perform writes
so we'd either need to pass this through the CheckpointManager or we'd
need a modified/another flushing task.

The simplest solution though is to simply stop counting prepares for
magma. We can maintain a correct document count for magma if we avoid
trying to count prepares. We can add a magma_dbdump and a prefix
range scan to count the number of prepares during a cbcollect to have
at least some diagnostics.

In a future patch I'll remove the MagmaDBStats prepare counting.

Change-Id: Iec97fc9a00d1837369644e5e845713f1276ab0b5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131656
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>