MB-36371: Use configuration.json purge age for couchstore
Currently when we compact a vBucket we use the ns_server provided
purge timestamp (a calculation based on the current time and the
metadata purge age) to determine if we should purge a tombstone. We
have the metadata purge age in the configuration already for magma
so switch over to using this for both.

In a future patch we can remove the API field.

Change-Id: I2142a920fe2c0415109077544eae4788b4869d86
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131808
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>