MB-40262: Subdoc inserts of Alive empty docs should be '{}'
Prior to the fix for MB-40126 (c55541f9f), if subdoc created a new
document (via doc_flag::Add or MkDoc) and all of the paths were XATTR
paths then the resulting document body would be '{}' - i.e. an empty
JSON object, but crucually *not* an empty value.

As part of fixing MB-40126 (subdoc mutations creating a docuement in
the *deleted* state had a non-empty '{}' value), the '{}' value for
alive document was lost - i.e. patch c55541f9f made both Alive and
Deleted inserts with only XATTR paths have a null '' value.

While in the abstract this is probably fine, it breaks the previous
5.0.0 -> 6.5.0 behaviour, and applications may be relying on this
(SDK Transactions does).

As such, revert to the previous behaviour for Alive documents - they
will be created as '{}' when all paths are XATTR paths.

(The previous fix for Deleted documents is kept - they will be created
as '').

Change-Id: I7b96b89a656c7b745fcd3c19174c6859e5c957f2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131817
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>