MB-38202: Ensure we atomically dec onDiskTotalItems after compaction
Currently in EPBucket::compactionCompletionCallback() we decrement the
number of items on disk by calling EPVBucket::getNumTotalItems() to get
the number of items on disk and the subtracting the value by
'ctx.stats.collectionsItemsPurged'. Then setting onDiskTotalItems to
this value using EPVBucket::setNumTotalItems().

However, this means there is a small window where the value of
onDiskTotalItems could be modified between our read and write of
onDiskTotalItems, which would cause us to have an inconsistent on disk
item count. To fix this we need to use std::atomic<> decrement
function. To do this, we've refactored VBucket::incrNumTotalItems() and
VBucket::decrNumTotalItems() to also take an arg of the delta of which
the onDiskTotalItems should be incremented or decremented. Which has the
default value of one.

Change-Id: I208da5141844dce99a8e5f8a722fb89e9e83f44d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131936
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>