MB-35071: Use a single worker thread in bucket test to avoid races
The test used multiple connections to the server, and the setup to
cause the operation to block was two commands. We could therefore
end up that the server would run the first command and not the
second before deleting the bucket. Given that the second command
wasn't executed yet (the one which would put the connection in
blocked state) the server would delete the bucket immediately
and the second command would be discarded (and connection deleted).

During cleanup of EWB it would now stop the threads and that would
cause the connection to be signalled (but the connection isn't valid
anymore because it never entered the ewb state).

Change-Id: I3fdec4c3e092b31889ca0bd8b57687aef80f8270
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132019
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>