MB-40365: Allow ById scan to resume from last key after a pause
A scan can pause and resume, in the case of this MB it was the backfill
buffer being full that triggered a pause. Prior to this fix, when the
scan resumes in the ById case the scan is started again, this can just
hit the same condition and the backfill just loops through pause/resume
indefinitely.

This commit adds a lastReadKey member to the ById context which is set
to the last read key when pause occurs, now the scan can be resumed
from the position in the by-id index it got to.

Change-Id: Idc4f1dd51af0c23cc957ae7e3cbabadf6e149a7d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132233
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>