MB-26334: Don't count 'system events' after warmup
When a system-event is flushed, we don't increment the vbucket's item
count. So create a collection and the item count won't change, however
after warmup the item count would increase.

This commit fixes this issue by subtracting the number of system events
from the item count when warming up. This works well for 7.0 as only
1 system event per collection exists and when a collection is deleted
the same system event is deleted (and the collection no longer exists
in the manifest).

Change-Id: Icf76753638370b36bfacd73b512716a8317039f1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132370
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>