MB-40415: Update local documents as part of commit
Previously compaction happened over two steps; first the database
was compacted and documents purged before the local documents
(_local/vbstate and _local/collections/dropped) was updated to
reflect the changes. This results in two different database
headers being written to the file. If the system crashed before
the last database header was written to disk we would end up
with an inconsistent vbucket file as we would use the compacted
file without the local documents updated.

This patch performs the modification in a precommit hook so that
we'll compact and update the local document within the same
couchstore commit.

Change-Id: I9e3d2de19ff08e595e4aae324f3c0b858f3afafd
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132420
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>