MB-39594: Re-factor CachingReadHandle construction
The getManifestEntry function shows up in the top-10 when profiling a
front-end thread under operation load. The more frequent use-case for
this method is with allowSystem=false (front-end thread), so rather than
having one path trying to support all cases, have two methods for each
use case.

This isn't expected to improve performance as we're only removing one
branch, but the simplified path is easier to reason about.

Change-Id: I0e7cbfff2053ec3011179838e1b87856a7a66c43
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132447
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>