MB-16181: Remove 'caching' of default collection from VB::Manifest
The VB::Manifest object 'cached' the presence of the default collection
in a boolean, this was orinially added to mitigate against the impact of
the developer preview code in the GA 6.5 release of Couchbase Server.
The idea was that we reduce or completley remove the need to do a map
search for the default collection items. This caching is now redundant
and has mostly been removed already (we have been doing map lookups for
all collections for some time now).

Change-Id: I2b7a6d73dc7ce6a845bee02f39ad3b634a3abe2b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132650
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>