MB-40467: Don't use updateStoredValue in VBucket::handlePreExpiry
At "expiration by access" currently we replace the SV's value by calling
HT::unlocked_updateStoredValue and then we execute the soft-delete
logic.

The problem with updateStoredValue is that the function updates the
deleted-state of the StoredValue. While that is fine for the usual
write-path, in the case of expiration we esentially loose the "pre
expiration deleted-state", which leads to errors in the accounting of
BasicLinkedList::numDeletedItems for Ephemeral. See the code in
BasicLinkedList::updateNumDeletedItems for details.

The problem is currently hidden by the main issue caught in MB-40467,
and an ep_testsuite test fails as soon as we fix MB-40467.
So, this patch is pre-requirememnt for the actual main fix for MB-40467.
The follow-up patch contains the fix for MB-40467 and the test coverage
for it. Plus, the existing ep_testsuite covers what we fix in this
patch.

Change-Id: Id5821f13f0c9d239ec368e03b7887611246f9b14
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132727
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>