MB-40517: Don't generate a VBCBAdaptor task name for invalid vbuckets
Spotted whilst investigating issue that we can log 'slow' messages for
invalid vbuckets (vb:1024 in this case). This diverted attention from
the actual issue.

Commit changes VBCBAdaptor so that the vbid we embed in the task name
only reflects real vbuckets the VBCBApdator was able to get a handle on.
We do this by using an invalid vbucket ID as the intialiser for the
class member 'currentvb' and only update it when the task locates a
vbucket.

In the MB logs this will mean we don't log vb:1024 anymore, but the last
visited vbucket and also means cbstats should only ever return task
names for real vbuckets owned by the node.

Change-Id: I664628b84af36e564a8f03bed0349a379f7a6d91
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132753
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>