MB-32358: Reduce the need for rollback for collection streams
Reduce the need of rollback for collection streams that are lagging
behind the vbucket high seqno. To do this add logic to the our rollback
code to ensure we don't force the client to rollback if they have seen
all mutations/events for the collections that they are streaming even if
the vbucket purge seqno is higher than the requested seqno e.i.
(client.start.seqno < vb.purge.seqno AND client.start >= collection.high
AND collection.high <= vb.purge). We also perform this check for a
second time between the point of which we decided if we're performing
rollback and if not when we then double check before performing
backfill in case the purge seqno has advanced between the stream request
being excepted and backfill starting.

Change-Id: I546c6e05340cc0c5d9521fa287253e91d12b7a58
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133977
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>