MB-40480: Compare seqno at VBucket::deletedOnDiskCbk
Currently at VBucket::deletedOnDiskCbk we check the revision seqno
of the item to determine if it is the latest version before removing
it from the HashTable post-persistence if it is also deleted. This
causes a race condition though when we have deleted prepares and aborts.
This race condition is as follows:
    1) Create and flush a prepare
    2) Create an abort for that prepare
    3) Run the flusher to the point that it has acquired the mutations
       from the CheckpointManager but not yet invoked the
       PersistenceCallback
    4) Create a deleted prepare with the same key
    5) Finish running the PersistenceCallback

At step 5 we would then compare the deleted flags and revision seqno of
the abort with those of the new prepare we created at step 4. These
values will be the same and at this point we will remove the new
prepare from the HashTable. This causes errors/exceptions later on
as we attempt to complete this prepare as we expect that prepares are
always resident in the HashTable.

Correct this by changing the revision seqno check to an actual seqno
check which should be different for any non meta items.

Change-Id: Icd498725fab94a0339e6677b4d14c98c5a196b8e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132951
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>