MB-40517: ExecutorPool::cancel must be able to call schedule
The task cancel path can call schedule, this certainly happens for tasks
which have the final reference to a VBucket. The task destructs and
in-turn destructs the VBucket, which wants to schedule a clean-up task.
Prior to this fix, a deadlock occurred because schedule was called
whilst the ExecutorPool mutex was held (because we were inside cancel).

This fixes a regression introduced by 4d160d19.

Change-Id: Ia0b71a5f96efdb4813d65e9f6fea83ef008b8d8e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133137
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>