MB-40580: Add --file arg to mctimings to render JSON data from its self
Add an argument --file to mctimings, allowing it to consume JSON data
from mctimings --JSON. As part of this work also ensure that when the
argument --JSON is used for multiple operations that we dump valid JSON.
As currently we dump multiple JSON objects in the same file stream.
Instead this should be an JSON array of JSON objects.

Change-Id: Iff7e1f57ce1a1a7685f448099b5d9fe7cea9d6e4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133739
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>