MB-40580: Fix and update rendering of histograms by mctimings
Fix histogram rendering to stop it skipping buckets which have no count
values, as this looses useful information for the user about the true
value of percentiles. The occurrence of buckets with no count is by
design due to the why HdrHistogram_c iterates though its internal data
structure.

Add a legend for the rendered histogram to clarify the meaning of
each part of the histogram.

Update the rendering of time values to use floating point values so we
can represent time values to 2dp which is a little more natural for a
user to read e.g.
Before:[   930 -   1090 ]us (99.9658%)     212|
After: [  0.93 -   1.09 ]ms (99.9658%)     212|

Change-Id: I042feffcb268bdefa5f84917cbd5f30c79d6ce47
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133733
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>