Delay calculating the max connections
As part of the refactor to move the calculation of the max
number of file descriptors available it started to print out
the new limit to the _console_ as part of reading the initial
settings.

To work around this problem don't install the "hook" to the
settings parser until after we've set up the logger and then
call it to make sure it ends up in our log file.

This removes lines like the follwing on the console every
time memcached is started (but puts it in memcached.log after
we print out the server version etc)

2020-07-28T09:20:04.562957+02:00 INFO recalculate_max_connections: max_fds:525310, max_connections:1000, system_connections:5000, engine_fds:523274

Change-Id: I5531d669b4d50a870a2a95649e1cdbcdd199352d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133349
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>