MB-39292: 1/n set_collections persist current manifest
set_collections only allows 'forward' progress after checking
that the new manifest is a successor of the current manifest,
however after warm-up we have to accept whatever we are given.

This commit updates set_collections so that for persistent buckets
the new manifest is first stored in the database directory and
then we update from that manifest on-success, now warm-up brings
the manifest back from storage and we can validate that further
updates are a valid successor.

Ephemeral buckets just update with no background task.

This patch stores using Manifest::toJSON and reloads JSON using
Manifest's existing construction with no integrity checking.

Change-Id: Ie548e31f56c4847ecf4c0c4ad866544f6bcd2a5c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137161
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>