MB-40830: Take VBucket state lock when setting manifest at warmup
At warmup completion we set the vBucket manifest as ns_server may send
us a new manifest before we create our vBuckets. This setting of the
manifest should only be done to active vBuckets and it does not take a
vBucket state lock which means that the state of the VBucket can change
while we set the manifest.

In the case of this MB this issue appears to manifest when we transition
from active->replica->active while dropping a collection due to manifest
application. This causes a brief window in which we can crash as we
attempt to notify a PDM of a dropped collection when it (the PDM) does
not exist. It's not enough to take the VBucket state lock only when
dropping the collection though as we should not be attempting to apply
a manifest to a replica vBucket in this way anyway (a replica vBucket
should only set the manifest via DCP SystemEvents).

Take the VBucket state lock when setting the manifest at warmup to
ensure that we only set the manifest on active vBuckets and to ensure
that the state does not change throughout the operation.

Change-Id: I3312766dce9379ca8d6e9c6510386aac55e94ae4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134134
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>