Set collections manifest on non-active -> active transition
Currently we set the collections manifest when transitioning from
replica to active. In a failover we can also transition from
pending to active, and need to set the manifest in this case too.
We should also consider the dead state though, we just want the vBucket
to be usable when we transition to active so we should also set the
manifest in this case too.

Change-Id: Ief42868f65239a4ba2172415e2872c5ee921b2ad
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134148
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>