cbstats tasks: Correctly flag 'running' tasks
The task state 'TASK_RUNNING' is misleading - it doesn't mean the task
is necessarily running, it could also be in the 'runnable' state -
i.e. ready to run but not yet assigned a CPU thread to execute
on. Moreover, due to the way it is used, a task which is currently
running, but has re-scheduled itself for it's _next_ execution will be
in the TASK_SNOOZED state.

This distinction is relevent because 'cbstats tasks' marks tasks which
are currently running with '*', and shows how long they have been
running for so far. However by simply checking 'state == TASK_RUNNING'
this incorrectly:

1. Includes tasks which are only Runnable (but not yet executing).
2. Exludes tasks which _are_ Running, but have snoozed themselves for next execution.

Fix this by only considering a task to be Running if it has a non-zero
'last_starttime'. This also changes the definition of last_starttime
inside ep-engine to be cleared (set to zero) when a task stops
executing.

Change-Id: I8dfadd2dbc90b6cbc998046c5a0281c09d265bba
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135700
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>