Fix doOneScopeStats() to handle invalid ScopeID's correctly
This patch fixes Collections::Manager::doOneScopeStats() by changing it
returns is scopeID after parsing a scopeID using std::stoi(). Currently
we create a cb::EngineErrorGetScopeIDResult() and instantiate using the
ScopeID() object. This is wrong, as this then cast implicitly to a
uint64_t, which is is treated as the manifestId and is thus assigned to
the manifestId member of cb::EngineErrorGetScopeIDResult() rather than
the scopeId member. Which could then end up with us perform incorrect
privilege checks for Collections::Manager::doOneScopeStats().

So to fix this ensure once we've successfully parsed the text to an int
we check that the scope exists in the manifest. Otherwise if not we
should return unknown_scope with the manifest uid. To perform the
scope validity test add method Collections::Manager::isScopeIDValid().

Change-Id: Ica2de8fced4d2d8b6be55eab35b2410125feddc4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134298
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>