MB-40718: Pending Prepare does not expire
TTL for SyncWrite is legal, but the document is supposed to expire if /
when it is committed.
Some paths in our code (eg, backfill) may access a logically-expired
Prepare and process the expire. That is wrong and may crash the
memcached process, as in the case of backfill we make the call in a
BG thread.

The fix in this patch ensures that the expire never happens on a
pending Prepare and that it does happen after the item is committed.

Change-Id: I23f5a4fca9ffca6d734b277969e3ea957ae80dee
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134356
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>