Build speed [5/6]: Reduce inclusion of nlohmann/json.hpp
Using ClangBuildAnalyzer highlighted that the header
'nlohmann/json.hpp' is the most expensive header to include when
building ep-engine_ep_unit_tests.

Overall time to build ep-engine_ep_unit_tests and top 5 expensive
headers:

Before:

      149.26 real      1422.24 user        81.96 sys

    *** Expensive headers:
    83770 ms: tlm/deps/json.exploded/include/nlohmann/json.hpp (included 135 times, avg 620 ms)
    74002 ms: ../kv_engine/engines/ep/src/vbucket.h (included 99 times, avg 747 ms)
    70453 ms: ../kv_engine/engines/ep/src/kv_bucket.h (included 67 times, avg 1051 ms)
    59612 ms: ../kv_engine/include/memcached/engine.h (included 150 times, avg 397 ms)
    59038 ms: ../kv_engine/engines/ep/src/collections/vbucket_manifest.h (included 46 times, avg 1283 ms)

json.hpp is both an expensive header to include (avg 620ms), and
included from many places. We cannot do much about the individual cost
(it's a third-party library), but we can reduce the number of places
it is included from.

One of the biggest contributors is the fact it is included from
vbucket.h (which is a very popular header) - this is because we pass
nhlohmann::json as an argument to some methods, but specifying a
default (empty) value. This requires the complete definition of the
type in vbucket.h - i.e. we cannot forward-declare:

    void setState(vbucket_state_t to,
                  const nlohmann::json& meta = {});

Essentially we have any optional `meta` argument, which is represented
by an empty JSON object. By changing the representation to a (possibly
null) pointer, we _can_ use a forward declaration:

    void setState(vbucket_state_t to,
                  const nlohmann::json* meta = nullptr);

Additionally, we previously used nlohmann::json directly for the
VBucket::replicationTopology member variable. Instead use pimpl to avoid
needing the definition in the vbucket.h

These changes result in a significant reduction in the cost of
nhlohmann/json.hpp itself, and vbucket.h - which go from 1st and 2nd
hottest headers respectively to 5th and 6th:

After
      148.18 real      1348.86 user        80.48 sys

    *** Expensive headers:
    60726 ms: ../kv_engine/engines/ep/src/collections/vbucket_manifest.h (included 48 times, avg 1265 ms)
    60532 ms: ../kv_engine/include/memcached/engine.h (included 150 times, avg 403 ms)
    60130 ms: ../kv_engine/engines/ep/tests/module_tests/evp_store_single_threaded_test.h (included 31 times, avg 1939 ms)
    59514 ms: ../kv_engine/include/memcached/dockey.h (included 209 times, avg 284 ms)
    55285 ms: tlm/deps/json.exploded/include/nlohmann/json.hpp (included 84 times, avg 658 ms)
    54296 ms: ../kv_engine/engines/ep/src/vbucket.h (included 99 times, avg 548 ms)

Change-Id: I39c409a7c4e20910df5c9e0ba34bf7cb5095d73d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134371
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>