MB-40135: Make CouchKVStore initialization easier to follow
Add some early returns and comments to the functions that we call when
we initialize a CouchKVStore to explain that should a vBucket file exist
without a vbucket_state local document then a default constructed
vbucket_state is used. This vbucket_state is dead by default and the
service should act as though the vBucket does not exist.

Add a unit test that demonstrates the above scenario too.

Change-Id: I4a9525003b4eb74eb26ee5ef537762c445c881bc
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134446
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>