MB-39505: Rework default labels in StatCollector
Previously, adding a default label modified the existing StatCollector,
and returned a handle which would remove the default label when it went
out of scope.

This might not be the expected behaviour as it does not follow common
patterns; with a "mutable" set of labels hidden in a collector, it is
not clear what a call to addStat will do.

Now, the new collector returned from `withLabels` can be used to add a
labelled stat, and the existing collector is not altered.

 StatCollector collector;
 auto labelledCollector =
     collector.withLabels({{"scope","scopeName"}});

 // add a stat without the scope label
 collector.addStat(...)

 // add a stat _with_ the scope label
 labelledCollector.addStat(...)

Change-Id: I400e9869517b45dbe0ee96d49748056af45e324c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134655
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>