Add kvlite skeleton
kvlite is a small program which offers a REST interface to the
cluster_test framework (if you want to do more manual testing /
experiments with memcached).

One of the "problems" if you want to mock around with vbucket
states on a normal installation of couchbase is that ns_server
monitors the state of the system and will try to make sure that
things are as they expect. kvlite will only start memcached
and allow you to create buckets, but once created it won't
try to monitor the vbuckets state or anything so you're
free to do whatever you want.

If you have kvlite running on a machine you can start use it
with:

    ./couchbase-cli cluster-init --cluster http://127.0.0.1:6666 \
                                 --cluster-username Administrator \
                                 --cluster-password asdfasdf \
                                 --cluster-ramsize 1024 \
                                 --cluster-port 6666 \
                                 --services data

kvlite will ignore the parameter you specify for ramsize and won't
stop you from creating buckets which eats up more than the memory
you've got on you system.

Once initialized you can create 30 buckets with:

    for f in `seq 30`
    do
     ./couchbase-cli bucket-create \
                     --cluster http://127.0.0.1:6666 \
                     --bucket mybucket-$f \
                     --bucket-type couchbase \
                     --storage-backend couchstore \
                     --durability-min-level none \
                     --bucket-ramsize 200 \
                     --bucket-replica 0 \
                     --bucket-priority low \
                     --bucket-eviction-policy fullEviction \
                     --enable-flush 0 \
                     --enable-index-replica 0 \
                     --compression-mode off \
                     -u Administrator \
                     -p asdfasdf
   done

You should be able to get the terse clustermap by requesting

    http://127.0.0.1:6666/pools/default/b/<bucketname>

Change-Id: I989127b1a05d7f6a7630c677bd92df2f092afaef
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133519
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>