MB-41024: Introduce MockReplicationThrottle
Add a mock subclass of ReplicationThrottle to allow tests to directly
control the behaviour of the ReplicationThrottle. This is particulary
useful when trying to test memory pressure scenarios - tests may want
the replication throttle to allow a mutation, but further down the
call chain have the VBucket method return ENOMEM. This is complex and
brittle to achieve by tweaking the mem_used / replication threshold
values directly.

Change-Id: Ifc530d535c47f75bc634dbf606b85765d217a6b2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134815
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>