MB-40382: Don't throw if we drop empty collection - magma
In magma-kvstore we throw if we attempt to drop an empty collection.
This code was added under the incorrect assumption that the only time
a collection can exist with a stat doc is if it's empty and it's the
default collection. This is not the case though. Any collection can
exist without a stat doc if it is empty. Remove this assertion from the
code and add a unit test for this case.

Change-Id: Ie52da4c0c950124f49f0ffe2765324d84527b083
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135051
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>