MB-41095: Use testPrivilege when checking DCP stream privileges
When re-evaluating DCP privileges we must not use checkPrivilege,
because on failure that function assumes it is being called from
the context of a incoming command, and will fail (as seen in the
MB) with an exception.

Using testPrivilege is much simpler and doesn't call functions
which may throw from this context.

Change-Id: I63d8b28b1fe6a93cfd144dc6967c3710389351c5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135367
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>