MB-36370: Optimize the cluster_testapp replication proxy
With regard to replication, a cluster_testapp test acts like a proxy
between the Producer/Consumer endpoints of a DCP connection.
Currently the proxy reads from the source socket and copies dato into
the destination socket. While that is perfectly fine for functional
tests, that introduces a relevant runtime overhead when testing DCP
performance as seen on tests
at http://review.couchbase.org/c/kv_engine/+/133499.

This patch modifies the proxy to using an enhanced bufferevent API that
moves data from a source socket to a destination socket and avoids any
unnecessary copy. In the DataCopy test used in MB-36370 we get +30% when
we run the test against a "fast consumer" and the old proxy becomes the
bottleneck.

Note that the enhanced proxy skips any packet-filtering, so we enable it
only if the user doesn't set any filtering function.

Change-Id: I050f92b7cbbdbd7a6c6345bc4b3732d4c527bec5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135423
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>