[BP] MB-41255: Don't bgfetch for a replica delete of an xattr
MB-36087 identified and fixed an issue with "del_with_meta" where
a delete against an evicted xattr crashed because we skipped doing
a bgfetch, KV crashed because the deleteWithMeta function went on
to try and prune the xattrs of the existing key, but had a null value.

An unexpected outcome of this fix is that DCP can now be exposed
to "would block" errors, this is because "del_with_meta" and DCP delete
share the same function.

MB-41255 demonstrates what happens when the consumer is told "would
block" for a delete. The DCP consumer waits for the bgfetch and then
retries the delete, but the consumer state machine was already moved
along by the initial delete, hence we see 'out of order' seqno errors
and the delete failing.

Note for backport: unit-test is rewritten as the original was using
test harness code that is not part of the alice branch

Change-Id: I6d9c52a0fca0ce7ab0cb201025c8f6057b681481
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136003
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>