MB-41283: Fix crash due to keyIndexes pointing to freed queued_items
Fix crashes and GSL preconditions being hit during checkpoint
processing. Due to the committedKeyIndex and metaKeyIndex point to
queued_items that have been freed. Due to the DCP consumer receiving a
disk snapshot without MARKER_FLAT_DISK set.

This can occur when the following sequence of events occur.
1. A Memory checkpoint is created this contains X mutations
   seqno:1 -> X with key:TEST at seqno:A being one of them.
   1. While parsing these mutations we insert key:Test into
      committedKeyIndex
2. The consumer receives a new disk snapshot seqno:X+1->Y with the
   MARKER_FLAG_CHK not set and containing a mutation for key:TEST at
   seqno:B
   1. We then update the original memory checkpoint to be a disk
      checkpoint
   2. Then we call Checkpoint::queueDirty() for key:TEST which then
      finds the original mutation at seqno:A we then remove the
      mutation at seqno:A from the toWrite queue of the checkpoint
      (which committedKeyIndex points to)
   3. We then FAIL to update the key:TEST in committedKeyIndex to point
      to seqno:B
3. Then another disk snapshot comes along without a MARKER_FLAG_CHK
   with a mutation for key:TEST at seqno:C
   1. We then go into Checkpoint::queueDirty() and find a key for TEST
   in committedKeyIndex but its now pointing to memory thats no longer
   valid as we've erased the qeued_item of seqno:A

To fix this ensure we still add items to the keyIndexes if the open
checkpoint was a memory checkpoint but has been updated to a disk
checkpoint. Otherwise still ensure we don't use the keyIndexes when we
have a disk checkpoint.

This patch also adds unit tests to ensure the steps above reproduce the
crashes seen and that this patch fixes the issue.

Change-Id: I7926d99c151dfb728d6d54c9966e614f106e6e6e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135630
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>